feat: extract SystemDisambiguationDialog to packages/ui with Storybook stories
Move the presentational component from desktop renderer to packages/ui
for Storybook coverage. Desktop re-exports from @gamelord/ui.

Stories cover: single file (2 systems), single file (3 systems),
multiple files with apply-to-all, mixed extensions, and long filenames.
