+++ packages/ui/index.ts
@@ -18,6 +18,7 @@ export * from "./components/GameLibrary";
 export * from "./components/ScrollLetterIndicator";
 export * from "./components/GameDetails";
 export * from "./components/ResumeGameDialog";
+export * from "./components/SystemDisambiguationDialog";
 export * from "./components/TVStatic";
 export * from "./components/PlatformIcon";
 export * from "./components/WebGLRenderer";
